Added Reference Environments
This are the environments I used while working on Blender, they won’t match 1:1 to the ones on Unity, since I got this ones from the Cinema4D files from Neil, but it would probably help to preview the cameras to begin with, and will help as reference to whoever is going to populate the world with the proper assets.

They are labeled by Scenes, as well as description.