Theatre staging
Had a meeting with Neil, and he explained me we should approach all the scenes based on 1 camera so that it takes less resources, and we can keep it more theatre like on the staging, so I am updating all the animation states, with simpler shots, all based on 1 camera, and this camera, has only animation when it needs to reframe a bit, or do a pan, but mostly the same camera all around.

Most poses are still placeholders just for staging, now I’m going to start working on their poses, to give them as much life as possible before we hit our deadline, and I move on into little people.